feat: Add Late Move Reductions (#15)
Passed STC:

Elo   | 23.93 +- 10.55 (95%)
SPRT  | 8.0+0.08s Threads=1 Hash=16MB
LLR   | 3.03 (-2.94, 2.94) [0.00, 5.00]
Games | N: 2676 W: 1043 L: 859 D: 774
Penta | [113, 248, 500, 296, 181]
https://openbench.chessnibble.com/test/119/

BENCH: 21337504

+++ src/search/bench.cpp
@@ -29,5 +29,8 @@ void search::bench::run(searcher& searcher, const u32 depth) {
     const auto seconds = elapsed / 1000;
 
     std::cout << std::format("\ninfo string {} seconds", seconds) << std::endl;
-    std::cout << std::format("{} nodes {} nps", total_nodes, total_nodes / seconds) << std::endl;
+    std::cout << std::format(
+        "{} nodes {} nps", total_nodes,
+        static_cast<u64>(static_cast<double>(total_nodes) / static_cast<double>(seconds)))
+              << std::endl;
 }
